refactor(base): consolidate MilliBase wiring under MilliCache\Base
Move per-site/network Manager wiring (was Admin/UI/Pages), migrations,
and the MilliBase Settings factory into a single MilliCache\Base
namespace. Each scope class now owns its Settings instance and defaults
directly, removing the Core\Settings factory layer. Core\ is left
containing only generic plugin infrastructure (Loader, Storage,
Updater); Base\ holds everything that interfaces with MilliBase.
