+++ src/Admin/UI.php
@@ -12,18 +12,16 @@
 
 namespace MilliCache\Admin;
 
-use MilliCache\Admin\UI\CacheActions;
-use MilliCache\Admin\UI\Pages;
-use MilliCache\Admin\UI\StatusBuilder;
+use MilliCache\Base;
 use MilliCache\Engine;
+use MilliCache\Admin\UI\CacheActions;
+use MilliCache\Admin\UI\StatusBuilder;
 
 ! defined( 'ABSPATH' ) && exit;
 
 /**
  * Constructs the shared dependencies (StatusBuilder, CacheActions) and
- * the per-site and (on multisite) Network Admin pages.
- *
- * Each Page registers its own MilliBase Manager from its constructor.
+ * the per-site and (on multisite) network MilliBase Managers.
  *
  * @package    MilliCache
  * @subpackage MilliCache/Admin
@@ -35,7 +33,7 @@ final class UI {
 	 * Initialize the UI and boot the MilliBase facade.
 	 *
 	 * Managers are created at plugin load time, so schema-derived defaults
-	 * are available immediately. Each Page's config closure defers
+	 * are available immediately. Each Manager's config closure defers
 	 * translation calls to `init`.
 	 *
 	 * @since 1.3.0
@@ -48,10 +46,10 @@ final class UI {
 		$status_builder = new StatusBuilder( $engine, $plugin_name, $version );
 		$cache_actions  = new CacheActions( $engine );
 
-		new Pages\Site( $plugin_name, $version, $status_builder, $cache_actions );
+		new Base\Site( $plugin_name, $version, $status_builder, $cache_actions );
 
 		if ( $engine->multisite()->is_enabled() ) {
-			new Pages\Network( $plugin_name, $version, $status_builder, $cache_actions );
+			new Base\Network( $plugin_name, $version, $status_builder, $cache_actions );
 		}
 	}
 }
